requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
google-generativeai==0.3.1
requests==2.31.0
aiohttp==3.9.1
psutil==5.9.7
